requirements.txt
Flask
Flask-SQLAlchemy
xhtml2pdf
qrcode[pil]
resend
python-dotenv
